+++ R/add_global_p.R
@@ -128,7 +128,7 @@ add_global_p.tbl_regression <- function(x,
         columns = "p.value",
         label = "**p-value**",
         hide = FALSE,
-        fmt_fun = x$inputs$pvalue_fun %||% getOption("gtsummary.pvalue_fun", default = style_pvalue)
+        fmt_fun = x$inputs$pvalue_fun %||% .get_deprecated_option("gtsummary.pvalue_fun", default = style_pvalue)
       )
   }
   # adding global p-values
@@ -216,7 +216,7 @@ add_global_p.tbl_uvregression <- function(x, type = NULL, include = everything()
         columns = "p.value",
         label = "**p-value**",
         hide = FALSE,
-        fmt_fun = x$inputs$pvalue_fun %||% getOption("gtsummary.pvalue_fun", default = style_pvalue)
+        fmt_fun = x$inputs$pvalue_fun %||% .get_deprecated_option("gtsummary.pvalue_fun", default = style_pvalue)
       )
   }
   # adding global p-values