+++ R/as_gt.R
@@ -100,7 +100,7 @@ as_gt <- function(x, include = everything(), return_calls = FALSE, ...,
   # taking each gt function call, concatenating them with %>% separating them
   gt_calls[include] %>%
     # adding default gt formatting options
-    c(parse_expr(getOption("gtsummary.as_gt.addl_cmds", default = "NULL"))) %>%
+    c(parse_expr(.get_deprecated_option("gtsummary.as_gt.addl_cmds", default = "NULL"))) %>%
     # removing NULL elements
     unlist() %>%
     compact() %>%