+++ R/tbl_survival.R
@@ -69,7 +69,7 @@ tbl_survival.survfit <- function(x, times = NULL, probs = NULL,
   # setting defaults -----------------------------------------------------------
   estimate_fun <-
     estimate_fun %||%
-    getOption("gtsummary.tbl_survival.estimate_fun") %||%
+    .get_deprecated_option("gtsummary.tbl_survival.estimate_fun") %||%
     switch(is.null(times) + 1,
       partial(style_percent, symbol = TRUE),
       partial(style_sigfig, digits = 3)